+++ sklearn/model_selection/__init__.py
@@ -3,6 +3,8 @@ from .partition import LeaveOneOut, LeavePOut, KFold, \
         ShuffleSplit, StratifiedShuffleSplit, train_test_split, check_cv
 from .validate import cross_val_score, permutation_test_score, \
         learning_curve, validation_curve
+from .search import GridSearchCV, RandomizedSearchCV
+from .utils import ParameterGrid, ParameterSampler, fit_grid_point
 
 __all__ = ['KFold',
            'LeaveOneLabelOut',
@@ -18,4 +20,9 @@ __all__ = ['KFold',
            'train_test_split',
            'learning_curve',
            'validation_curve',
+           'GridSearchCV',
+           'RandomizedSearchCV',
+           'ParameterGrid',
+           'ParameterSampler',
+           'fit_grid_point',
           ]
